split storage.rs into modules
